+++ database/migrations/20-v_deviation_papperskorgar_gavle.sql
@@ -29,8 +29,8 @@ ALTER TABLE upstream.mv_match_papperskorgar_gavle OWNER TO app;
 
 CREATE OR REPLACE VIEW upstream.v_deviation_papperskorgar_gavle AS
 	SELECT
-		31 AS dataset_id,
-		16 AS layer_id,
+		31::bigint AS dataset_id,
+		16::bigint AS layer_id,
 		upstream_item_ids,
 		CASE
 			WHEN osm_element_id IS NULL THEN upstream_geom
